backend/finder_backend/requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.1
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
ecdsa==0.19.1
gunicorn==23.0.0
idna==3.10
packaging==25.0
psycopg2-binary==2.9.10
pyasn1==0.4.8
PyJWT==2.10.1
python-jose==3.4.0
requests==2.32.3
rsa==4.9.1
six==1.17.0
sqlparse==0.5.3
urllib3==2.4.0
